+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Current
 
-TODO
+- Add default HTTP request retry policy to `Client`
+- Add default HTTP user agent to `Client`
+- More direct data feed infrastructure
 
 # 0.11.1