+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Current
 
+- TODO
+- 
+# 0.12
+
 - Add default HTTP request retry policy to `Client`
 - Add default HTTP user agent to `Client`
 - More direct data feed infrastructure